We can use a box-shadow with a 5px spread to imitate a 5px border
Since there's a border-radius applied, we can't just apply a border to
the image. And no pseudo elements on images means that won't work. Was
using a <b> since it has no semantic meaning in HTML5, but even better
is to just use box-shadow with a 5px spread so we don't need an extra
element.