Remove custom title color and link color
Unfortunately it's the law that `wp_head()` has to be the absolutely
last thing before the closing head tag. Since `print_scripts` runs
before `enqueue_scripts`, there's no way to override the styles in the
stylesheet. I'm not going to use !important on this because it breaks
too much stuff and I'm not going to load styles in the footer becuase
that's dumb.